sermon_monitor/requirements.txt
yt-dlp
google-generativeai
python-dotenv
schedule
